1-strdup.c

+++ 0x0B-malloc_free/1-strdup.c
@@ -21,7 +21,7 @@ char *_strdup(char *str)
 
 	else
 	{
-		for (i = 0; i < sizeof(str); i++)
+		for (i = 0; i < (int)sizeof(str); i++)
 			s[i] = str[i];
 	}
